tests works local, now testing travis

+++ src/main/java/entities/RenameMe.java
@@ -12,19 +12,19 @@ import javax.persistence.NamedQuery;
 //@NamedQuery(name = "RenameMe.deleteAllRows", query = "DELETE from RenameMe")
 public class RenameMe implements Serializable {
 
-    private static final int serialVersionUID = (int) 1L;
+    private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     
     public RenameMe() {
     }
         
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
